datapath: add tail call hooks to run programs for custom metrics
Add tail call hooks to bpf_lxc in order to run custom programs. The
objective is to attach user-specified programs in order to collect
custom metrics.

Three tail calls are added, one on the egress path, the other two for
distinct branches on the ingress path. The custom program is passed
through dedicated, per-endpoint prog_array maps. These maps are created
by tc when loading bpf_lxc programs, and managed by the Cilium agent
afterwards. Each prog_array map has two entries, one for ingress and one
for egress. On the ingress path, make sure we skip the tail call when
the packt is being redirected to a L7 proxy, to avoid running the custom
program twice on the incoming packet.

In this commit, tail call is attempted unconditionally. Follow-up commit
will guard it with a macro so it can be removed at compilation time.

On failure to jump through the tail call, the behaviour of the bpf_lxc
program is mostly unchanged. On success, it jumps to a custom program
and never goes back to the original program. The hook is inserted at the
very end of the original program, to avoid skipping packet processing
operations. However, to preserve a consistent behaviour, the return
value from the final eBPF programs must remain the same. To that end,
the return value is passed through the cb buffer so the custom program
can eventually return it.

At this stage, custom programs have to be loaded, attached, and
referenced in the prog_array maps by the user (typically with clang and
bpftool).

Custom tail calls are not supported for host endpoints at this stage.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>