datapath: add tail call hooks for custom programs to IPv6 datapath
On the same model that the hooks added to the IPv4 datapath, add hooks
for tail calls to user-defined eBPF programs to provide a way to
implement, for example, custom metrics collection for IPv6 flows.

The prog_array maps holding the references to the custom programs are
extended with two new cells, for IPv6 ingress and egress paths.

Signed-off-by: Quentin Monnet <quentin@isovalent.com>