cilium: encryption, allow programming IPSec FWD policy and template IPs
[ upstream commit d02b6b74447335abe36c6b9b59027c5417d647ee ]

So far we've programmed FWD policy rules to mirror the IN policy rules,
but this requires the outer IPSec tunnel dstIP and inner dstIP cidr to
be non-overlapping. Also we've been setting the template IPs the same
as the matching src/dst IPs. We are about to fix the ENI case where these
assumptions are no longer true.

In order to support this allow pushing unique FWD/IN policies and
also specify matching IPs and template IPs.

Note: It probably makes sense to do a heavier refactor here and rip
out UpsertIPsecEndpoint(), at this point it just adds confusion in
my opinion. But, lets avoid that at the moment for backports.

Signed-off-by: John Fastabend <john.fastabend@gmail.com>